BUG: Fix hardware.profiles examples
The change to use common cloud engine code caused the TAC and native key lookup engines to return data where the hardware.profiles property was an AspectPropertyValue instance but the individual properties (e.g. hardware.properties[0].ismobile) were not.
This was unintended and caused the TAC and native key lookup examples to fail. This engine has now been modified to use a custom processInternal function to handle this correctly.